add TaskSettingsInterface
