+++ src/ibl_mesoscope_to_nwb/mesoscope2025/datainterfaces/__init__.py
@@ -18,7 +18,7 @@ from ._meso_wheel_interfaces import (
     MesoscopeWheelMovementsInterface,
     MesoscopeWheelPositionInterface,
 )
-from ._meso_session_epoch_interface import SessionEpochsInterface
+from ._task_settings_interface import TaskSettingsInterface
 from ._visual_stimulus_interface import VisualStimulusInterface
 
 __all__ = [
@@ -34,6 +34,6 @@ __all__ = [
     "MesoscopeWheelMovementsInterface",
     "MesoscopeWheelPositionInterface",
     "MesoscopeDAQInterface",
-    "SessionEpochsInterface",
+    "TaskSettingsInterface",
     "VisualStimulusInterface",
 ]
